Improve integration with Facebook
Prior to this, in certain cases, e.g. using the test apps with a FB
account not authorized to use the apps, the website doesn't load
at all. The reason is that Angular was bootstrapped after the
initialization of FB SDK, so if there was a problem in the latter,
Angular wouldn't be bootstrapped.