Implement a service for authentication
To be able to identify the currently logged in user any where, along
with information about the user, I implemented a service that contains
information like FB user ID, FB access token, user permission, etc.
This service can be injected anywhere to know information about
the user. As a test, it was used in the hadithtags page to show or
hide the delete button depending on whether the user has the the
necessary permission or not.